+++ packages/mirage-time.2.0.0/unix/time.ml
@@ -1,7 +1,5 @@
 (*
- * Copyright (c) 2011-2015 Anil Madhavapeddy <anil@recoil.org>
- * Copyright (c) 2013-2015 Thomas Gazagnaire <thomas@gazagnaire.org>
- * Copyright (c) 2013      Citrix Systems Inc
+ * Copyright (c) 2017 Thomas Gazagnaire <thomas@gazagnaire.org>
  *
  * Permission to use, copy, modify, and distribute this software for any
  * purpose with or without fee is hereby granted, provided that the above
@@ -16,14 +14,4 @@
  * OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
  *)
 
-type error = [`Unimplemented | `Disconnected]
-
-let pp_error ppf = function
-  | `Unimplemented -> Fmt.string ppf "operation not yet implemented"
-  | `Disconnected  -> Fmt.string ppf "device is disconnected"
-
-module type S = sig
-  type +'a io
-  type t
-  val disconnect: t -> unit io
-end
+let sleep_ns ns = Lwt_unix.sleep (Duration.to_f ns)