+++ packages/mirage-random.2.0.0/CHANGES.md
@@ -1,3 +1,9 @@
+### v2.0.0 (2019-10-21)
+
+* Specialise buffer to Cstruct.t directly in Mirage_random.S (#12 @hannesm)
+* Deprecate Mirage_random.C (#12 @hannesm)
+* Raise lower bound of OCaml to 4.06.0 (#12 @hannesm)
+
 ### v1.2.0 (2018-11-11)
 
 * Move Stdlibrandom to a separate package, mirage-random-stdlib